Move updater support scripts to lib directory
