+++ updater/lib/tinglesoftware/dependabot/clients/azure.rb
