+++ updater/lib/tinglesoftware/dependabot/vulnerabilities.rb
